Use outcome_same_sex_marriage_and_civil_partnership_not_possible
Instead of
outcome_same_sex_marriage_and_civil_partnership_in_other_countries.

These two outcomes are identical so I've replaced the use of one with
the other. I'll remove the unused outcome in another commit.

I chose to use the outcome with the "not_possible" suffix as I think in
general we want outcome names that describe the result rather than the
inputs.