Introduce on_response blocks
TODO: Add tests

The idea is that we'll use these to set attributes on the calculator/policy
object instead of doing that in `next_node` blocks.

One advantage of doing this is to separate out the attribute-setting code which
should make it easier to do automatically in future.

Another advantage is that we can avoid having to pass the response into any
validation-related methods on the calculator which should make them more
ActiveModel-like.

The new `on_response` blocks are run after `next_node_calculation` blocks, but
before `validate` blocks. Ideally I would've preferred to have them run *before*
`next_node_calculation` blocks too, but since the `calculator` is currently
usually instantiated in a `next_node_calculation` bloc, this seemed like the
simplest solution.

An alternative would be to use a `precalculate` block to instantiate the
`calculator`, however, currently these don't get executed for the first
question.