Use attribute instead of param in validation methods
This makes the validations closer to the ActiveModel style.

TODO: Avoid the slight bodge in `#valid_last_payday_before_offset?` - I had to
add 1 day to the RHS to counteract the fact that the validation was previously
done on the raw response rather than `relevant_period_from` which is the raw
response plus 1 day.