Define Marriage outcomes for Dominican Republic
This commit contains changes to both opposite and same sex marriages in the
Dominican Republic. This outcome seeks to isolate the outcomes for opposite
and same sex from other due to the increased paths and possible future changes
that could occur.

Here it starts with an if statement for the page title, followed by the body tag.
Within the body tag decisions are the made based on the rules given to the
developer.

This also contains the outcome_dominican_republic and redirects the flow for all
marriage outcomes where the country of ceremony is the Dominican Republic

This commit adds logic that displayed information for a UK resident/user
planning on getting married in the Dominican republic.
Previous message remains the same for the other cases

Updates have been made the Dominican Republic fees table utilizing the new
the_consular_fees_table partial.

Replace consular address with an appointment link for same sex marriages in
Dominican Republic

Finally, content updates and changes to text for Dominican Republic