Move Register a Birth fees into rates file
This will allow us to add rates for 2016/17 that can be deployed before
6 April 2016.

I wanted to set the start_date as 2015-04-06 but that causes the
register-a-birth regression tests to fail after I add rates for 2016/17.
This is because the date for the regression tests is hardcoded as 1 Jan
2015 which means that no suitable rates are found, which means that the
`RatesQuery` falls back to using the most recent rates (2016/17).
Setting the start_date to 2015-01-01 ensures that these rates continue
to be used in the regression tests.