Use on_response blocks in check-uk-visa
It seems better to separate the setting of attributes on the calculator from
the routing logic in `next_node` blocks. The recently introduced `on_response`
block allows us to do this.

Note that I've had to set the `calculator` state variable in the first
`on_response` block, because these blocks are executed before
`next_node_calculation` blocks.