Introduce SmartAnswer::Block
This does pretty much everything that `SmartAnswer::Calculation` does, except
that it doesn't assign the result of the block to a specified state variable.
Hopefully you can see this by comparing the unit tests for the two classes.

I'm planning to use this to introduce a new type of block to the Flow/Question
DSL which will be called immediately after receiving the response from the
user. However, I've named the new class very generically, because there's no
reason it couldn't be used at other points in the DSL.

I did contemplate other slightly more specific names e.g.
`NonAssigningCalculation` and `BlockEvaluatableInContextOfState`.

For the former, I thought that the `Calculation` suffix was a bit confusing,
because I think it conveys the idea that it has a result.

I thought the latter was unwieldy and came to the conclusion that really this
is just a block - it's just enhanced with a method that allows it to be called
in the context of a `SmartAnswer::State`. And the method signature of the
`#evaluate` method is enough to explain the latter.

Thus I think naming it `SmartAnswer::Block` is the best option.