Introduce on_response blocks to flow/question DSL
The idea is that we'll use these blocks to set attributes on the
calculator/policy object instead of doing that in `next_node` blocks.

One advantage of doing this is to separate out the attribute-setting code which
should make it easier to do automatically in future.

Another advantage is that we can avoid having to pass the response into any
validation-related methods on the calculator which should make them more
ActiveModel-like.

The new `on_response` blocks are run *before* both `next_node_calculation` and
`validate` blocks.

The existing unit tests around this area of the codebase leave a bit to be
desired, but I've added new tests as best I can roughly following existing
patterns. I did also contemplate adding an `EngineIntegrationTest` with a
new fixture flow, but I decided that it would be overkill and added the test to
`FlowTest` instead. I feel as if this gives sufficient test coverage for the
new functionality.