Use on_response blocks in part-year-profit-tax-credits
It seems better to separate the setting of attributes on the calculator from
the routing logic in `next_node` blocks. The recently introduced `on_response`
block allows us to do this. It also means the response will be available on the
calculator as an attribute when validation methods are called which means we
won't need to pass the response into these methods, thus simplifying the code.

Note that I've had to set the `calculator` state variable in the first
`on_response` block, because these blocks are executed before
`next_node_calculation` blocks.