Simplify validation methods in calculate-statutory-sick-pay
Now that the responses are available as attributes we don't need to pass the
response into the validation methods. This makes the validations closer to the
ActiveModel style.

I've had to use a slight bodge in `#valid_last_payday_before_offset?`. I've
added 1 day to the RHS to counteract the fact that the validation was previously
done on the raw response rather than `relevant_period_from` which is the raw
response plus 1 day. I'm planning to improve this later as part of this Trello
card [1].

[1]: https://trello.com/c/uu0HeZee