Handle template exceptions when generating indexable_content
I came across this problem when trying to run the `panopticon:register` rake
task locally as suggested by @chrisroos.

Unfortunately the `MethodMissingHelper` & `MethodMissingObject` approach, which
we use to render node templates in the absence of any state variables, isn't
perfect.

In particular in the `outcome_os_other_countries` outcome in `marriage-abroad`
there is a `case` statement which switches on `calculator.ceremony_country`.
When rendering this outcome for `indexable_content` purposes, this will be an
instance of `MethodMissingObject` and will not match *any* of the `when`
clauses. Thus execution will fall through to the `else` statement which raises
an exception.

I couldn't see a good way to avoid this happening and so I decided to rescue
exceptions raised when rendering a template. This means that we won't get any
indexable content for the affected node(s). However, I don't think this is too
bad since only one node is currently affected.

Note that we have to render the whole template to extract any one piece of
content (e.g. title, body, etc), so if there an exception is raised in any one
part of the template, then we won't be able to extract any of the pieces of
content.

I've also added an integration test which builds a `FlowRegistrationPresenter`
for each flow and calls the `indexable_content` method on each of them. This
replicates the essentials of the `panopticon:register` rake task which is
called on deployment, although it does actually not send the indexable content
to Panopticon.

Although this test takes 15 seconds to run on my local machine I think it's
worth having to avoid possible errors on deployment.

Note also that the rescuing of `ActionView::Template::Error` only applies in
the context of the generation of indexable content at deploy time i.e. such
exceptions will not be rescued during user requests in production.