Handle template exceptions when generating indexable_content
I came across this problem when trying to run the `panopticon:register` rake
task locally as suggested by @chrisroos.

Unfortunately the `MethodMissingHelper` & `MethodMissingObject` approach, which
we use to render node templates in the absence of any state variables, isn't
perfect.

In particular in the `outcome_os_other_countries` outcome in `marriage-abroad`
there is a `case` statement which switches on `calculator.ceremony_country`.
When rendering this outcome for `indexable_content` purposes, this will be an
instance of `MethodMissingObject` and will not match *any* of the `when`
clauses. Thus execution will fall through to the `else` statement which raises
an exception.

I couldn't see a good way to avoid this happening and so I decided to rescue
exceptions raised when rendering a template. This means that we won't get any
indexable content for the affected node(s). However, I don't think this is too
bad since only one node is currently affected.

Note that we have to render the whole template to extract any one piece of
content (e.g. title, body, etc), so if there an exception is raised in any one
part of the template, then we won't be able to extract any of the pieces of
content.