Re-start A/B/C test on 'Start now' buttons
This commit reverts https://github.com/alphagov/smart-answers/pull/2337.
We ran a multivariate test for 'start now' buttons on 3 start pages in early March 2016. We want to run the test again to increase our confidence that the improvement trends shown in the original test are valid, with the same options winning and with a similar scale if difference.