Remove :administering_oath_or_making_declaration
Trello card: https://trello.com/c/qMeYVdfh

I searched the marriage-abroad regression test artefacts for
"Administering an oath or making a declaration". It appears in the
results for a number of different ceremony countries. I picked Albania,
Finland and Luxembourg at random and found those in the spreadsheet. It
would appear, based on the services in the spreadsheet, that this should
be removed rather than replaced with an alternative.

I've updated the regression test artefacts given that this is a
relatively small change.