Replace :affirmation_for_marriage
Trello card: https://trello.com/c/qMeYVdfh

Replaced with :affidavit_or_affirmation_for_marriage.

I searched for "Affirmation for marriage" in the marriage-abroad
regression test artefacts. I found a match for Finland. Checking the
spreadsheet for that country suggests that it should now be using
"Affidavit or affirmation for marriage" instead of "Affidavit for
marriage".

I've updated the regression test artefacts given that this is a
relatively small change.