Replace :affirmation_for_freedom_to_marry_in_english_and_chinese
Trello card: https://trello.com/c/qMeYVdfh

Replaced with :affidavit_or_affirmation_for_marriage.

I searched for "Affirmation for freedom to marry in English & Chinese"
in the marriage-abroad regression test artefacts. I found a match for
China. Checking the spreadsheet for that country suggests that it should
now be using "Affidavit or affirmation for marriage" instead.

I've updated the regression test artefacts given that this is a
relatively small change.