Remove :making_an_oath
Trello card: https://trello.com/c/qMeYVdfh

I searched the marriage-abroad regression test artefacts for "Making an
oath". It appears in the results for Egypt.  Checking the spreadsheet
for this country suggests that it should now be using "Affidavit or
affirmation for marriage" instead. That service is already being
included in the place that previously rendered the consular_fees_table
partial with :making_an_oath.

I've updated the regression test artefacts given that this is a
relatively small change.