Remove :affirmation_for_freedom_to_marry_in_other_language
Trello card: https://trello.com/c/qMeYVdfh

I searched the marriage-abroad regression test artefacts for
"Affirmation for freedom to marry in any other language". It appears in
the results for Lebanon, Philippines and United Arab Emirates.  Checking
the spreadsheet for those countries suggests that they should now be
using "Affidavit or affirmation for marriage" instead. That service is
already being included in both places that rendered the
consular_fees_table partial with
:affirmation_for_freedom_to_marry_in_other_language.

I've updated the regression test artefacts given that this is a
relatively small change.