Remove :issuing_letter_or_certificate_english
Trello card: https://trello.com/c/qMeYVdfh

I searched the marriage-abroad regression test artefacts for "Issuing
any other consular letter or certificate in English". It appears in the
results for a number of different ceremony countries. I picked Albania
and Kyrgyzstan at random and found those in the spreadsheet.  It would
appear, based on the services in the spreadsheet, that this should be
removed rather than replaced with an alternative.

I've updated the regression test artefacts given that this is a
relatively small change.