Replace :affirmation_or_affidavit_for_freedom_to_marry
Trello card: https://trello.com/c/qMeYVdfh

Replaced with :affidavit_or_affirmation_for_marriage.

I searched for "Affirmation for freedom to marry" in the marriage-abroad
regression test artefacts. I found a match for Peru. Checking the
spreadsheet for that country suggests that it should now be using
"Affidavit or affirmation for marriage" instead.

I've updated the regression test artefacts given that this is a
relatively small change.