Update `minimum_hourly_rate` with the living wage value.
After 1 April 2016, if you are over 25 you will get 7.20 as the
Minimum Living Wage, so the calculator should be updated accordingly.

Test cases should be updated because the test data is heavily based 
on 25 years old to exercise the National Minimum Wages rates. 
Hence this commit updates them to 24 years old, and create a new test 
case for 25.