Add links to Panopticon and router-data repos
1. Add a link to the Panopticon repository

The first step in the archiving documentation mentions the Panopticon application,
but does not give any details about what Panopticon is or does. As I don't
understand what Panopticon does yet, I've a link to the repository (and the
accompanying README).

2. Change router-data "command" into a link to the router-data repository

The second step in the archiving documentation makes router-data look like a command
that needs to be run. I changed the format to a link to the repository to make it clearer
that this is something that needs to be updated and deployed.