Add MarriageAbroadCalculator#services
This returns the most appropriate services for the combination of
`ceremony_country`, `sex_of_your_partner`, `resident_of` and
`partner_nationality`.

Both `ceremony_country` and `sex_of_your_partner` are required. It's possible
to define services for values of `resident_of` and `partner_nationality` by
setting them to `default` in the data. This avoids us having to duplicate data
where the same services are used for all opposite_sex services, for example.

I've updated the _consular_fees_table partial to read from
calculator.services and to fall back to the `services` passed in the
call to `render`. This should allow me to incrementally move services
from the templates to the new marriage_abroad_services.yml file.

I'd really like a unit test around the behaviour of _consular_fees_table
but can't see an easy way to make that work.

The marriage-abroad regression tests are passing at this commit (albeit
the checksum data is out of date).