Revert "TEMP: Update marriage-abroad-responses-and-expected-results.yml"
This reverts commit 2e47b23bf13ce9a70f1cc02ef99cf6d9a061a19e.

I was relying on having this complete set of responses in order to be able to
reliably use the marriage-abroad-countries-leading-to-outcome.rb and
marriage-abroad-countries-leading-to-outcome-for-country.rb scripts to extract
the service information from the templates. Now that all the services have been
extracted, this file can be reverted to its previous state.