Update opposite sex services for affidavit or affirmation countries
Trello card: https://trello.com/c/D0ZMM9v5

I manually copied the data from the spreadsheet attached to the Trello card to
this data file in preparation for updating these services.

Updated using:

    $ rails r script/update-marriage-abroad-services.rb \
    script/marriage-abroad-services-updates/opposite_sex_affidavit_or_affirmation.yml

Note that this change implies that we should show the default services for
ceremonies in Norway and Slovenia where the user is resident of a
third_country. At the moment, logic in the flow/template means this isn't the
case. I plan to address this separately.

Note that although this has added a 'opposite_sex' section for colombia, it
won't be shown due to logic in the flow/templates. I plan to address this
separately.