Update opposite sex services for certificate of custom and law countries
Trello card: https://trello.com/c/D0ZMM9v5

I manually copied the data from the spreadsheet attached to the Trello card to
this data file in preparation for updating these services.

Updated using:

    $ rails r script/update-marriage-abroad-services.rb \
    script/marriage-abroad-services-updates/opposite_sex_certificate_of_custom_and_law.yml

Note that despite the addition of this data, none of these services will be
shown due to logic in the flows/templates.  I plan to address this separately.
This also explains the lack of changes to the regression test artefacts after
this change.