Update opposite sex services for CNI countries
Trello card: https://trello.com/c/D0ZMM9v5

I manually copied the data from the spreadsheet attached to the Trello card to
this data file in preparation for updating these services.

Updated using:

    $ rails r script/update-marriage-abroad-services.rb \
    script/marriage-abroad-services-updates/opposite_sex_cni.yml

Note that although this has added a 'opposite_sex' section for ukraine, it
won't be shown due to logic in the flow/templates. I plan to address this
separately.