Update same sex services for Spain
Trello card: https://trello.com/c/D0ZMM9v5

I manually copied the data from the spreadsheet attached to the Trello card to
this data file in preparation for updating these services.

Updated using:

    $ rails r script/update-marriage-abroad-services.rb \
    script/marriage-abroad-services-updates/same_sex_spain.yml

I also manually removed the residency specific services that were there.  The
addition of the default set of services means that we no longer need these
residency specific services.

Note that this change implies that we should show the default services for
ceremonies where the user is resident of a third_country. At the moment, logic
in the flow/template means this isn't the case. I plan to address this
separately.