Remove unused marriage-abroad services
Trello card: https://trello.com/c/S9dKYBuT

I searched for each of the services in _consular_fees_table_items and removed
those that no longer appear in marriage_abroad_services.yml.

I checked the resulting list of services with the list in the linked Trello
card and am happy that we're only left with those that we're expecting.