Update start date for marriage-abroad rates
This is to keep the regression tests happy. The regression tests are run
with a fixed time of 1 Jan 2015. If the RatesQuery can't find some
matching rates then it'll use the most recent. This means that when we
add 2016/17 rates the regression tests will start using those instead of
these earlier rates.