Revert artefacts for `minimum-wage-calculator-employers`"
PR #2423 included an issue when calculating the minimum wage for past 
years if the user is over 25. The returned value is £7.20, which is the 
minimum living wage, and it should be calculated according to the
the yaml file.

This commit revert the artefacts as a firs step to fix this issue.
This way we can start fresh, so any change on the artefacts can be
checked against a working version.