Update `national_living_wage_rate` in calculator
Previous implementation had the value hardcoded to 7.2
This commit uses the `national_hourly_rate` method because they both share
the same values in the yaml file.

This will be valid until October. So a more robust solution should
be put in place. This improvement would be out of the scope of this PR
because of a strict deadline.