Update `minimum_hourly_rate` to rely on yaml file
`minimum_hourly_rate` in the calculator relies on the yaml file to
do calculations. The reason is: after 2015-10-01 the national minimum
wage and the national living wage share the same values. 
This could change, though, in October.

The yaml file with the rates have been update to calculate the 
minimum living wage for users over 25.

Due to the strict deadline no further refactoring is done.