Fix minimum wage rates for people aged 24
Trello card: https://trello.com/c/n72PekVB

This fixes a problem with am-i-getting-minimum-wage when the user gives
their age as 24. The implementation of
`MinimumWageCalculator#per_hour_minimum_wage` means that no data was
being returned, causing a "NoMethodError: undefined method `[]' for
nil:NilClass" exception to be raised.

I added the tests to expose the error before updating the data in the
YAML file.