Update marriage_abroad_services for Ukraine
Trello card: https://trello.com/c/2jG2LmZi

This data implied that the services would be shown to residents of Ukraine.
We've now had confirmation that this shouldn't be the case. I've updated the
fee data to make it explicit that they're not shown for ceremonies in Ukraine.

This change isn't strictly necessary as logic in the flow/templates meant that
we weren't showing these services. I've updated because I think it's a good
idea to keep this data as accurate as possible.