Update marriage_abroad_services for Norway
Trello card: https://trello.com/c/xQGZ5uuL

This data implied that the services would be shown to residents of a
third_country. We've now had confirmation that this shouldn't be the case. I've
updated the fee data to make it explicit that they're only shown for residents
of the UK and Norway.

This change isn't strictly necessary as logic in the flow/templates meant that
we weren't showing these services. I've updated because I think it's a good
idea to keep this data as accurate as possible.