Update update-worldwide-location-organisation-data.rb
I've updated the script so that it updates organisation data for all locations
stored in worldwide_locations.yml.

It's not obvious to me that it's useful to update organisation data for a
single location (although it might well be).

I've updated the script so that it stores "pretty" JSON. This makes it easier
to read the fixture data (if necessary) and will make it easier to see any
changes that are made to the fixture data when we run this script in future.