Avoid fixture data in '.for_location' test
I want to be able to update the fixture data (it's also used by the regression
tests) without having to update these tests.

I've replaced the "finding organisations in a location" test with an
alternative that tests the `.for_location` method by stubbing the
`GdsApi::Worldwide#organisations_for_world_location` method. The previous
implementation used WebMock to stub the real API and relied on the behaviour of
`#organisations_for_world_location`.