Avoid fixture data in delegation tests
I want to be able to update the fixture data (it's also used by the regression
tests) without having to update these tests.

I'm using an `OpenStruct` for the organisation data as that's what we get from
`GdsApi::Worldwide#organisations_for_world_location`, which is used to
instantiate `WorldwideOrganisation`s in `WorldwideOrganisation.for_location`.