Add integration test for am-i-getting-minimum-wage
Add an integration test to exercise code for users over 25 under
after 2016-04-06.

Artefacts and test-unit integration tests use 2015-01-01 as the
reference date; this approach prevents the tests from exercising all
branches for the outcomes: `current_payment_above` and
`current_payment_below` because the National Living Wage only applies
after 2016-04-06.

This test was added to increase the test coverage of the existing
flow.