squash! Handle template exceptions when generating indexable_content
Add an integration test which builds a `FlowRegistrationPresenter` for each
flow and calls the `indexable_content` method on each of them. This replicates
the essentials of the `panopticon:register` rake task which is called on
deployment, although it does actually not send the indexable content to
Panopticon.

Although this test takes 15 seconds to run on my local machine I think it's
worth having to avoid possible errors on deployment.