Update legalisation text for ceremonies in sweden when resident of uk
I determined that it was the
outcome_opposite_sex_marriage_in_consular_cni_countries_when_residing_in_uk_or_ceremony_country
we needed to amend by looking at the responses-and-expected-results.yml file
for marriage-abroad ceremonies in Sweden.

The `###Legalisation and translation` header, and `You should check if your CNI
needs to be:` text normally comes from the "legisation_and_translation_intro_uk"
partial. This text is slightly different to that in the partial. We no longer
want the "You might need to exchange..." text, and "You should also check if it
needs to be:" has changed to "You should check if your CNI needs to be:".

* TODO: We should check whether the "You might need to exchange..." text should
still apply to other countries that lead to this outcome.

* TODO: We should check whether the "You should check if your CNI needs to be:"
text can be used for all ceremony countries that lead to this outcome.

Update regression test artefacts for marriage-abroad