Use partials in outcome_ceremonies_in_sweden_when_residing_in_sweden
There may be more partials we can use in this template but this is probably
enough to demonstrate the approach.

It's best to use `--ignore-all-space` when viewing the differences in the test
artefacts.

TODO: Question whether the differences in the test artefacts are OK. If not,
then can we update the partial so that it brings all uses of the partial into
line?