Fix Style/SymbolProc violations
I chose to fix some of these violations manually, because I wasn't happy with
the changes generated by Rubocop's auto-correct mechanism.

I also chose to disable the cop for a couple of the violations, because I
couldn't see a way to satisfy the cop without making the code much less
readable. The code in question is pretty horrible and will hopefully go away
when we get round to refactoring these flows. I feel that disabling the cops in
these cases just serves to highlight the problem code and should encourage us
to fix it!