Nest benefit-cap-calculator data under 'default' key
This was cherry-picked from 5fc48d741d6d96a6f59f28473ae3db184806694e. I've
removed the addition of data under the `future` key but will add that in a
later commit.

Use the `--ignore-all-space` option to better see the changes in this commit.

The benefit-cap-calculator regression tests are passing so I've updated the
associated checksum data.