Add a new first question to benefit-cap-calculator
TODO: Consider adding only single response to the first question for now.

To ask the user whether they'd like to check whether their benefits will be
capped based on the current rules, or the new rules which are going to come
into effect in Autumn 2016.

The question and template have been cherry-picked from
60303ff972ef46c468dd4c57f0ee57a1ee0a2fb0. That commit contained a breaking
change to the flow which hasn't been included here.

This commit has no real effect at the moment, as we'll give the same advice
irrespective of the answer the user has given.

I've updated the benefit-cap-calculator integration test:

* I've added a test for the new question

* I've nested all the existing tests under a new "when using pre-Autumn 2016
  rates and rules" context. Use the `--ignore-all-space` option to better see
  this change.

* I've duplicated all the existing tests under a new "when using post-Autumn
  2016 rates and rules". I think these are going to come in useful as the
  behaviour between the two paths starts to diverge.

I've manually added options for the new question to
benefit-cap-calculator-questions-and-responses.yml and then regenerated
benefit-cap-calculator-responses-and-expected-results.yml.

I've updated the regression test artefacts in this commit. The changes include:

* Renaming existing artefacts from /benefit-cap-calculator/ to
  /benefit-cap-calculator/default.

* Added new artefacts for /benefit-cap-calculator/future.

* You can see that the _outcome_ artefacts are currently identical for /default
  and /future by running `diff -r
  test/artefacts/benefit-cap-calculator/{default,future}`.

* You can see that the question page artefacts are all in the /default
  namespace by running `diff -r
  test/artefacts/benefit-cap-calculator/{default,future}`.

Given that I'm happy that the regression test artefact changes are all valid,
I've also updated the checksum data as part of this commit.