Remove worldwide_locations.yml from checksum data
We were inconsistently tracking worldwide_locations.yml in the checksum data.
The regression tests for every Smart Answer that contains a `country_select`
will be affected if the countries in worldwide_locations.yml change (at the
very least the html of the question page including the `country_select` will
have changed), yet we're only tracking it for three Smart Answers. Rather than
tracking worldwide_locations.yml in all Smart Answers that contain a
`country_select` I think it might be easier to rely on the documentation in
updating-worldwide-fixture-data[1], which explains that you'll need to run all
integration and regression tests after updating this data.

[1]: https://github.com/alphagov/smart-answers/blob/master/doc/updating-worldwide-fixture-data.md