Remove organisation fixtures from checksum data
We were inconsistently tracking organisation fixtures in the checksum data.
The regression test artefacts for several Smart Answers that contain
`country_select`s and display organisation data will be affected if the related
organisation fixture data changes, yet we're only tracking it for four Smart
Answers. Rather than trying to determine which Smart Answers require which
organistion fixtures I think it might be easier to rely on the documentation in
updating-worldwide-fixture-data[1], which explains that you'll need to run all
integration and regression tests after updating this data.

[1]: https://github.com/alphagov/smart-answers/blob/master/doc/updating-worldwide-fixture-data.md