Introduce new routing question
It only has a single option that routes to the standard set of
questions. I'm adding this as a first step as it'll hopefully be easier
to see the changes introduced by the 'future' benefit caps.

I've updated questions-and-responses so that it has an answer to the
first question.

I've updated responses-and-expected-results based on this change.

I've regenerated regression test artefacts. They should all be renames
(moving from top-level to 'current' directory).