Add duplicate questions for Q4 and Q6 for the future flow
Users are asked to select which flow they want to traverse in the new question 1.

However there isn't an easyish way to store the result from the question in a variable
that can be accessed by the whole Flow class.  The chosen flow value can only be accessed
in next_node blocks etc.  To get around this limitation we have duplicated Q4 and Q6 to be
able to hard-code the value of the flow we need to use.