DRY up benefit-cap-calculator shared logic
This commit makes use of the newly introduced ability to specify local variables
to be supplied in the evaluation of shared logic. We can now use a single
parameterized shared logic file for both pre-2016 and post-2016, i.e. we can
now delete the `benefit-cap-calculator-post-2016` shared logic.

The `node_name` lambda is used to add the '-post-2016' suffix when appropriate.
There's probably a simpler/better way to do this, but this was the best I could
come up with in a short time. `node_name` needs to be a lambda and not a method
in order that we have access to the `node_suffix` local variable.