Remove carers and guardian from 2016 benefits list
The `receiving_non_exemption_benefits_post_2016` symbolic link has been replaced
with a real template with the two benefits removed from the options.

The `receiving_exemption_benefits_post_2016` symbolic link has been replaced
with a real template with the two benefits added to the bullet point list.

The `carers_amount` & `guardian_amount` questions are now only added for
pre-2016. Thus their post-2016 templates can be removed.

Note that the "routing" to the `carers_amount` & `guardian_amount` questions is
still in place, but I wanted to avoid cluttering up the code with more
conditionals and, since the options no longer exist, these `when` statements
will never be exercised in post-2016 scenarios.

You might find it clearer to see what's going on in this commit if you use the
`--ignore-all-space` to view the diff.