Delete the oldest period (Oct 2008 - Sep 2009) from past payments
BIS responded to the factcheck agree that we only need to go back 6 years.

The template for minimum-wage-calculator-employers also included text for years going back to 2005. As these dates can't be
selected by the user (they do not appear in the past_payment question in the minimum wage shared logic)
they have been removed from the template.

Remove check for 2008-2009 from the integration tests.