Extract payment info from outcome_opposite_sex_marriage_in_consular_cni_countries_when_residing_in_uk_or_ceremony_country
I determined the list of countries that can lead to this outcome by using:

    $ rails r script/marriage-abroad-countries-leading-to-outcome.rb \
    outcome_opposite_sex_marriage_in_consular_cni_countries_when_residing_in_uk_or_ceremony_country

I've updated the add-payment-partials-to-marriage-abroad-services-data.rb
script to extract the payment info for these countries.

The output of running the script was:

Entry not found for afghanistan.
Entry not found for american-samoa.
Entry not found for andorra.
Entry not found for benin.
Entry not found for bhutan.
Entry not found for brazil.
Entry not found for burkina-faso.
Entry not found for cape-verde.
Entry not found for central-african-republic.
Entry not found for chad.
Entry not found for comoros.
Entry not found for congo.
Entry not found for djibouti.
Entry not found for equatorial-guinea.
Entry not found for eritrea.
Entry not found for gabon.
Entry not found for guinea.
Entry not found for guinea-bissau.
Entry not found for haiti.
Entry not found for iraq.
Entry not found for libya.
Entry not found for liechtenstein.
Entry not found for mali.
Entry not found for marshall-islands.
Entry not found for mauritania.
Entry not found for micronesia.
Entry not found for niger.
Payment method partial already set for norway. Existing value: pay_by_visas_or_mastercard. New value: pay_by_cash_or_credit_card_no_cheque.
Entry not found for palau.
Entry not found for paraguay.
Entry not found for rwanda.
Entry not found for san-marino.
Entry not found for sao-tome-and-principe.
Entry not found for south-sudan.
Entry not found for suriname.
Entry not found for timor-leste.
Entry not found for togo.
Entry not found for western-sahara.

I've confirmed that the changes the regression test artefacts are all expected
based on the output from running the script. The payment information has been
removed from all those listed, apart from Norway. The payment information for
Norway has changed. I plan to follow up on this to work out what the payment
information should be.